Move the decomposed-value walk into the registry
An `UnfoldPlan` says which leaves a value is delivered as and which hoists they
are reached through. Walking it — binding each hoist once, reaching each leaf
through its own path, deciding at every step whether what is held is owned or
borrowed — was JniGen's code, over the registry's plan.

The walk is the registry's now. `prebindgen_registry::unfold::walk` holds
`bind_hoists`, `reach_leaf_flat`, `Hoisted` and the four step helpers they are
built from, generic over one small trait: `DecomposedLeaf` is the five facts
the walk needs about a leaf — its path, its name, whether it is the value
itself, the reading it delivers, and whether the last step is a field read.
JniGen's `OutWire` implements it and keeps everything else it carries: wire
type, projection, jvalue layout, descriptors.

Nothing about the emitted code changes: every generated artifact is
byte-for-byte identical, and `covertest-kotlin` passes its 61 sections on the
JVM.

`emit/delivery.rs` loses 414 lines. What stays there is JNI's: the leaf
encoder, the builder and fold calls, local-frame sizing, cached method lookup,
exception routing — and `reach_leaf`, the optional-step-handling variant that
belongs with the delivery it serves until step 5 decides what to do with it.

Part of #596.
